Fuse sync install and start fragment into a single one
