+++ app/src/main/res/layout/fragment_sync_install.xml
@@ -62,6 +62,7 @@
             android:padding="@dimen/marginStandardMedium">
 
             <TextView
+                android:id="@+id/installDescription"
                 style="@style/Body.Secondary"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
@@ -87,14 +88,38 @@
                     android:layout_height="match_parent"
                     android:layout_weight="1" />
 
-                <com.google.android.material.button.MaterialButton
-                    android:id="@+id/installButton"
-                    style="@style/TextButtonPrimaryMedium"
+                <FrameLayout
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="@string/buttonInstall" />
+                    android:layout_height="wrap_content">
 
+                    <com.google.android.material.button.MaterialButton
+                        android:id="@+id/installButton"
+                        style="@style/TextButtonPrimaryMedium"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="@string/buttonInstall" />
+
+                    <TextView
+                        android:id="@+id/installedTextView"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:text="Installée"
+                        android:visibility="gone" />
+
+                </FrameLayout>
             </LinearLayout>
         </LinearLayout>
     </androidx.cardview.widget.CardView>
+
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/startButton"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/buttonStartSync"
+        android:visibility="gone"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
